Create WebUI type and separate web code.
Create web.ui for web ui type and it's handling functions.

Add reloadsig.go and reloadsig_windows to allow reparsing templates on SIGUSR1.
